Moved stuff to Shared AutoPausedField now
