+++ Content.Server/Animals/Components/EggLayerComponent.cs
@@ -2,7 +2,6 @@ using Content.Server.Animals.Systems;
 using Content.Shared.Storage;
 using Robust.Shared.Audio;
 using Robust.Shared.Prototypes;
-using Robust.Shared.Serialization.TypeSerializers.Implementations.Custom;
 
 namespace Content.Server.Animals.Components;
 
@@ -11,13 +10,13 @@ namespace Content.Server.Animals.Components;
 ///     It also grants an action to players who are controlling these entities, allowing them to do it manually.
 /// </summary>
 
-[RegisterComponent, Access(typeof(EggLayerSystem))]
+[RegisterComponent, Access(typeof(EggLayerSystem)), AutoGenerateComponentPause]
 public sealed partial class EggLayerComponent : Component
 {
     /// <summary>
     ///     The item that gets laid/spawned, retrieved from animal prototype.
     /// </summary>
-    [DataField(required: true), ViewVariables(VVAccess.ReadWrite)]
+    [DataField(required: true)]
     public List<EntitySpawnEntry> EggSpawn = default!;
 
     /// <summary>
@@ -32,19 +31,19 @@ public sealed partial class EggLayerComponent : Component
     /// <summary>
     ///     Minimum cooldown used for the automatic egg laying.
     /// </summary>
-    [DataField, ViewVariables(VVAccess.ReadWrite)]
+    [DataField]
     public float EggLayCooldownMin = 45f;
 
     /// <summary>
     ///     Maximum cooldown used for the automatic egg laying.
     /// </summary>
-    [DataField, ViewVariables(VVAccess.ReadWrite)]
+    [DataField]
     public float EggLayCooldownMax = 90f;
 
     /// <summary>
     ///     The amount of nutrient consumed on update.
     /// </summary>
-    [DataField, ViewVariables(VVAccess.ReadWrite)]
+    [DataField]
     public float HungerUsage = 25f;
 
     [DataField] public EntityUid? Action;
@@ -52,6 +51,6 @@ public sealed partial class EggLayerComponent : Component
     /// <summary>
     ///     When to next try to produce.
     /// </summary>
-    [DataField(customTypeSerializer: typeof(TimeOffsetSerializer)), ViewVariables(VVAccess.ReadWrite)]
-    public TimeSpan NextGrowth = TimeSpan.FromSeconds(0);
+    [AutoPausedField]
+    public TimeSpan NextGrowth = TimeSpan.Zero;
 }
