+++ Content.Shared/Nutrition/SharedIngestionEvents.cs
@@ -1,34 +1,36 @@
-using Content.Shared.Nutrition;
-
-namespace Content.Server.Nutrition;
+namespace Content.Shared.Nutrition;
 
 /// <summary>
 ///     Raised directed at the consumer when attempting to ingest something.
 /// </summary>
-public sealed class IngestionAttemptEvent : SharedIngestionAttemptEvent
+public abstract class SharedIngestionAttemptEvent : CancellableEntityEventArgs
 {
     /// <summary>
     ///     The equipment that is blocking consumption. Should only be non-null if the event was canceled.
     /// </summary>
+    public EntityUid? Blocker = null;
 }
+
 /// <summary>
 /// Raised directed at the food after finishing eating a food before it's deleted.
 /// Cancel this if you want to do something special before a food is deleted.
 /// </summary>
-public sealed class BeforeFullyEatenEvent : SharedBeforeFullyEatenEvent
+public abstract class SharedBeforeFullyEatenEvent : CancellableEntityEventArgs
 {
     /// <summary>
     /// The person that ate the food.
     /// </summary>
+    public EntityUid User;
 }
 
 /// <summary>
 /// Raised directed at the food being sliced before it's deleted.
 /// Cancel this if you want to do something special before a food is deleted.
 /// </summary>
-public sealed class BeforeFullySlicedEvent : SharedBeforeFullySlicedEvent
+public abstract class SharedBeforeFullySlicedEvent : CancellableEntityEventArgs
 {
     /// <summary>
     /// The person slicing the food.
     /// </summary>
+    public EntityUid User;
 }
